+++ packages/plugin-transform-modules-commonjs/src/util/interop.ts
